add: normalizeDate + move parseDate and formatDiff
