+++ src/client/components/DiscordTSButton/formatTS.ts
@@ -1,5 +1,7 @@
-export function formatTS(time: number | number[]): string {
-  const [start, end = 0] = Array.isArray(time) ? time : [time];
+import { normalizeDate } from "src/time/normalizeDate";
+
+export function formatTS(time: number | [number, number]): string {
+  const [start, end] = normalizeDate(time);
   const text = `Starts: <t:${start}:F> (<t:${start}:R>)${
     end > 0 ? `\nEnds: <t:${end}:F> (<t:${end}:R>)` : ""
   }`;
