+++ src/client/components/TimeDisplay/DisplayDelta.tsx
@@ -2,7 +2,7 @@ import { useStore } from "@nanostores/react";
 import spacetime from "spacetime";
 import { intervalStore } from "src/client/stores/intervalStore";
 import { NoSSR } from "src/client/components/NoSSR";
-import { formatDiff } from "src/utils/formatDiff";
+import { formatDiff } from "src/time/formatDiff";
 
 interface DisplayDeltaProps {
   time: number;
