+++ src/time/formatDiff.ts
